added utils to i18n-copy
