+++ src/cli/i18n-copy/index.ts
@@ -1,75 +1,7 @@
 const fs = require("fs");
 import config from "../../config";
 import path from "path";
-import * as prettier from "prettier";
-
-interface Translation {
-  l: string;
-  v: string;
-}
-
-interface CustomerOverride {
-  customer: string;
-  tr: Translation[];
-}
-
-interface Message {
-  code: string;
-  tr: Translation[];
-  overrides?: CustomerOverride[];
-}
-
-//checked single file
-const checkFile = async (inputPath: string, from: string, to: string) => {
-  const fileContent = require(inputPath);
-  let addValue = "";
-  let existLang = false;
-
-  const newContent = fileContent.messages.map((messages: Message) => {
-    messages.tr.forEach((value: Translation) => {
-      if (value.l === from) {
-        addValue = value.v;
-      }
-      if (value.l === to) {
-        existLang = true;
-      }
-    });
-    if (!addValue) {
-      throw new Error("--from " + from + " not exist");
-    }
-    if (!existLang) {
-      messages.tr.push({ l: to, v: addValue });
-    }
-
-    if (messages.overrides) {
-      messages.overrides.forEach((i: CustomerOverride, index) => {
-        let addOverrideValue = "";
-        let existOverridesLang = false;
-        i.tr.forEach((value: Translation) => {
-          if (value.l === from) {
-            addOverrideValue = value.v;
-          }
-          if (value.l === to) {
-            existOverridesLang = true;
-          }
-        });
-
-        if (!addOverrideValue) {
-          throw new Error("--from " + from + " not exist");
-        }
-
-        if (!existOverridesLang) {
-          messages.overrides![index].tr.push({ l: to, v: addOverrideValue });
-        }
-      });
-    }
-
-    return messages;
-  });
-  const writeContent = "exports.messages = " + JSON.stringify(newContent);
-  const prettierContent = await prettier.format(writeContent, { parser: "babel" });
-  fs.writeFileSync(inputPath, prettierContent, "utf-8");
-};
+import { checkMultipleFile, checkSingleFile } from "./utils";
 
 const langCopy = async () => {
   try {
@@ -84,17 +16,18 @@ const langCopy = async () => {
     }
 
     if (isDir) {
-      fs.readdirSync(inputPath).map(async (fileName: any) => {
+      const filesCheck = fs.readdirSync(inputPath);
+
+      for (const fileName of filesCheck) {
         if (fileName !== "index.js") {
           const filePath = path.join(inputPath, fileName);
-          await checkFile(filePath, from, to);
+          await checkMultipleFile(filePath, from, to);
         }
-      });
-      console.log("Successfully copied");
+      }
     } else {
-      await checkFile(inputPath, from, to);
-      console.log("Successfully copied");
+      await checkSingleFile(inputPath, from, to);
     }
+    console.log("Successfully copied");
     return 0;
   } catch (e) {
     console.error(e);
